requirements.txt
flask
joblib
flask-cors
python-dotenv
supabase
